+++ src/parser/DASHTree.h
@@ -45,7 +45,7 @@ public:
   virtual void SetLastUpdated(std::chrono::system_clock::time_point tm){};
   void SetUpdateInterval(uint32_t interval) { updateInterval_ = interval; };
   uint64_t pts_helper_, timeline_time_;
-  uint32_t firstStartNumber_;
+  uint64_t firstStartNumber_;
   std::string current_playready_wrmheader_;
   std::string mpd_url_;
